concurrent mail expunge and mail read
sdbox-mail uses sdbox_mail_set_expunged method to mark a mail as
expunged, if it no longer exists on the filesystem. I adapted this
behavior to rbox implementation. If a mail read (stat,read,getxattr)
operation fails due to object does not exist error ENOENT(-2), it is
assumed that the mail was deleted but current process did not know about
this. Dovecot handles this situation without throwing an error.