#106: missing check if connection to rados exists when cleaning up
in case connection to rados can not be opened in save_finished,
r_ctx->failed is set to true and  rbox_transaction_save_rollback is
called. transaction_save_rollback then tries in every case to remove
already writen objects or object parts from rados. This fails if the 
connection is closed. 
FIX: before trying to remove objects, a check is performed if connection
is open.