fix: failing unit test (dovecot_master)
due to additional checks of mail_storage_settings, it is required to set
the setting postmaster_address to a valid string, else validation will
fail.
Currently the workaround is to modify the mail_storage_settings
attribute postmaster_address. Unfortunatelly we can only get a const
reference to mail_storage_settings so we need to cast_const it away and
change the value in postmaster_address.