#150: repair index generate uid if, index no longer exist.
in case the index has been deleted, and user tries to repair the mailbox
the original mailbox name stored in attribute 'B' is used to group
objects into a mailbox. This could lead to redundant uids (e.g. through
copying / moving mails without update of the 'B' attribute). In this
case the uids is generated starting from 1 and all mails with the given
mailbox name ('B') are added to the mailbox.